+++ src/app/components/FooterMenu/FooterMenu.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import HomeIcon from "../Icons/HomeIcon";
+import ProfileIcon from "../Icons/ProfileIcon";
 import SavedDetailsIcon from "../Icons/SavedDetailsIcon";
 import SearchFieldIcon from "../Icons/SearchFieldIcon";
 import styles from "./FooterMenu.module.css";
@@ -9,8 +9,8 @@ function FooterMenu(): JSX.Element {
   return (
     <div className={styles.footerMenu}>
       <div className={styles.navBar}>
-        <Link to="/">
-          <HomeIcon />
+        <Link to="/profile">
+          <ProfileIcon />
         </Link>
         <Link to="/search">
           <SearchFieldIcon />
