+++ src/app/pages/Login/Login.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import LoginLogo from "../../components/Icons/LoginLogo";
 import MainBubble from "../../components/Styling Elements/MainBubble";
-
 import styles from "./Login.module.css";
 
 function Login(): JSX.Element {
@@ -11,7 +10,7 @@ function Login(): JSX.Element {
         <div className={styles.logo}>
           <LoginLogo />
         </div>
-        <a className={styles.loginButton} href="/githublogin">
+        <a className={styles.loginButton} href="/oauth/githublogin">
           Log in with GitHub
         </a>
       </div>
